addressed more comments
