+++ templates/login_or_signup.tmpl
@@ -29,13 +29,13 @@
             <div class="mui-row">
                 <div class="mui-col-sm-6 mui-col-md-4 mui-col-sm-offset-3 mui-col-md-offset-4">
                     <ul class="mui-tabs__bar mui-tabs__bar--justified">
-                        <li id="signup-tab">
+                        <li>
                             <a data-mui-toggle="tab" data-mui-controls="signup-form">
                                 Signup
                             </a>
                         </li>
 
-                        <li id="login-tab" class="mui--is-active">
+                        <li class="mui--is-active">
                             <a data-mui-toggle="tab" data-mui-controls="login-form">
                                 Login
                             </a>
